@@ -9,12 +9,12 @@
   },
   "dependencies": {
     "acorn": "^7.1.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-redux": "^7.1.3",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.4.1",
-    "redux": "^4.0.5",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0",
+    "redux": "^4.1.2",
     "redux-saga": "^1.1.3"
   },
   "devDependencies": {
@@ -25,6 +25,9 @@
     "puppeteer": "^13.0.1",
     "redux-saga-test-plan": "^4.0.4",
     "serve": "^13.0.2"
+  },
+  "jest": {
+    "testEnvironment": "jsdom"
   },
   "browserslist": {
     "production": [
