@@ -1,17 +1,17 @@
 import React from 'react';
 import {
 	BrowserRouter as Router,
-	Switch,
-	Route
+	Route,
+	Routes
 } from 'react-router-dom';
 import { ErrorNotFoundComponent } from '../components/error-component/error-component';
 import { IndexPage } from './index-page/index-page';
 
 export const Navigator = () => (
 	<Router>
-		<Switch>
-			<Route exact path='/' component={IndexPage}/>
-			<Route component={ErrorNotFoundComponent}/>
-		</Switch>
+		<Routes>
+			<Route exact path='/' element={<IndexPage/>}/>
+			<Route element={<ErrorNotFoundComponent/>}/>
+		</Routes>
 	</Router>
 );
